pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "insta-bot-gemini"
version = "1.0.0"
description = "Instagram DM bot powered by Google Gemini AI"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["instagram", "bot", "gemini", "ai", "chatbot"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
]

dependencies = [
    "flask>=2.3.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "pylint>=2.15",
]

[project.urls]
Homepage = "https://github.com/yourusername/insta-bot-gemini"
Repository = "https://github.com/yourusername/insta-bot-gemini.git"
Issues = "https://github.com/yourusername/insta-bot-gemini/issues"

[project.scripts]
insta-bot = "insta_bot.cli:cli"
